+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add new permissions modal component. Refs UIU-629, UIU-631.
 * Update `<PaneFooter>`: support arbitrary rendering of the content on two sides. Refs STCOM-521.
 * Update integration tests for new MCL. Refs STCOM-363.
+* Don't fail a proxy relationship due to an empty account expiration date. Fixes UIU-820.
 
 ## [2.24.1](https://github.com/folio-org/ui-users/tree/v2.24.1) (2019-07-26)
 [Full Changelog](https://github.com/folio-org/ui-users/compare/v2.24.0...v2.24.1)